[Jinqiang] - cast the seek time as unit milliseconds to int type

+++ android/src/main/java/com/brentvatne/react/ReactVideoViewManager.java
@@ -131,7 +131,7 @@ public class ReactVideoViewManager extends SimpleViewManager<ReactVideoView> {
 
     @ReactProp(name = PROP_SEEK)
     public void setSeek(final ReactVideoView videoView, final float seek) {
-        videoView.seekTo(seek * 1000.0f);
+        videoView.seekTo((int)(seek * 1000.0f));
     }
 
     @ReactProp(name = PROP_RATE)
